+++ spine-ts/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 4.3.12 - 2026-07-23
+
 ### spine-core
 
 - Expose restore WebGL logic on ManagedWebGLRenderingContext, so applications that pass an existing WebGL context instead of a canvas can restore all registered WebGL resources.